+++ Changelog.md
@@ -10,7 +10,7 @@
 -   SD card info into the "Info" Menue (thanks to [@Slider007]( https://github.com/Slider0007))
 -   Added a logging of the values in a text table in `/log/data` - each measurement is one line
     -   Format: tabulator separated
-    -   Content: time, raw-value, return-value, pre-value, error-text, cnn-digital, cnn-analog
+    -   Content: time, name-of-number, raw-value, return-value, pre-value, change-rate, change-absolute, error-text, cnn-digital, cnn-analog
     -   ATTENTION: format not fully fixed yet!